requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
pika==1.3.2
alembic==1.15.2
python-dotenv==1.1.0
gunicorn==23.0.0
black==25.1.0
isort==6.0.1
flake8==7.2.0
pylint==3.3.6
mypy==1.15.0
sentry-sdk==2.26.1
newrelic==10.9.0
pytz==2025.2
passlib[bcrypt]==1.7.4
marshmallow==4.0.0
desert==2022.9.22
requests==2.32.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
requests-mock==1.12.1